apps/spu-ui: moving queue elements to the queue component
